Issue with a video in a div
I think you have a bug here.  When i have <div><video></video></div> I wasn't getting to the video.  It was finding the video but then there was no image tag so it would perform the HtmlText which we don't want.  So I think it the video or img tags should be in an or with contains then negated then ored with the negated hasContent.  I have only run a brief test on it.